backend/requirements.txt
Flask_SocketIO
flask_cors 
mysql-connector-python